+++ packages/sit-onyx/src/components/OnyxSelectDialog/OnyxSelectDialog.stories.ts
@@ -4,7 +4,7 @@ import OnyxSelectDialog from "./OnyxSelectDialog.vue";
 
 /**
  * Pre-built dialog where the user can select one value from a given list of options.
- * This component is e.g. internally used by the [OnyxColorSchemeDialog](/story/navigation-modules-colorschemedialog--docs).
+ * This component is e.g. internally used by the [OnyxColorSchemeDialog](/docs/navigation-navbar-modules-colorschemedialog--docs).
  *
  * ### Keyboard support
  *